refact: extrair mensagens de chat de skill para skill-chat-message.mjs (Fase 3 — Commit 2/4)
